@@ -38,8 +38,8 @@
   },
   "devDependencies": {
     "cypress": "^12.17.4",
-    "eslint": "^8.46.0",
-    "eslint-config-next": "^13.4.9",
+    "eslint": "^8.47.0",
+    "eslint-config-next": "^13.4.19",
     "eslint-plugin-cypress": "^2.14.0",
     "start-server-and-test": "^2.0.0"
   },
