@@ -4,31 +4,29 @@ import {
   withStyles,
   Grid,
   Button,
-  IconButton,
 } from "@material-ui/core/";
 import { configureAdapterWithNotification } from "../helpers/serviceMeshes";
-import { withSnackbar } from "notistack";
-import CloseIcon from "@material-ui/icons/Close";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import { updateAdaptersInfo, updateProgress } from "../../../lib/store";
+import { useNotification } from "../../../utils/hooks/useNotification";
 
 
 const styles = () => ({})
 
 const ServiceMeshConfig = ({
-  classes, adapterLoc, updateProgress, enqueueSnackbar, closeSnackbar, updateAdaptersInfo
+  adapterLoc, updateProgress, closeSnackbar, updateAdaptersInfo
 }) => {
-
+  const { notify } = useNotification()
   const handleAdapterConfigure = () => {
-    configureAdapterWithNotification(enqueueSnackbar, updateProgress, ServiceMeshConfigNotificationAction(closeSnackbar), adapterLoc, updateAdaptersInfo)
+    configureAdapterWithNotification(notify, updateProgress, adapterLoc, updateAdaptersInfo)
   }
 
-  const ServiceMeshConfigNotificationAction = (closeSnackbar) => (key) => (
-    <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
-      <CloseIcon />
-    </IconButton>
-  )
+  // const ServiceMeshConfigNotificationAction = (closeSnackbar) => (key) => (
+  //   <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
+  //     <CloseIcon />
+  //   </IconButton>
+  // )
   return (
     <>
       <Grid item xs={12} style={{ padding : "1rem" }}>
@@ -51,4 +49,4 @@ const ServiceMeshConfig = ({
 const mapDispatchToProps = (dispatch) => ({ updateProgress : bindActionCreators(updateProgress, dispatch),
   updateAdaptersInfo : bindActionCreators(updateAdaptersInfo, dispatch), });
 
-export default connect(null, mapDispatchToProps)(withStyles(styles)(withSnackbar(ServiceMeshConfig)))
+export default connect(null, mapDispatchToProps)(withStyles(styles)(ServiceMeshConfig))
