@@ -1,6 +1,7 @@
 /* eslint-disable react/display-name */
 import { IconButton } from "@material-ui/core"
 import CloseIcon from "@material-ui/icons/Close";
+import { EVENT_TYPES } from "../../../lib/event-types";
 
 export const closeButtonForSnackbarAction = (closeSnackbar) => (key) => (
   <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
@@ -8,18 +9,14 @@ export const closeButtonForSnackbarAction = (closeSnackbar) => (key) => (
   </IconButton>
 )
 
-export const successHandlerGenerator = (snackbar, action, msg, cb) => (res) => {
+export const successHandlerGenerator = (notify, msg, cb) => (res) => {
   if (res !== undefined) {
     if (cb !== undefined) cb(res)
-    snackbar(msg, { variant : "success",
-      action,
-      autoHideDuration : 3000, })
+    notify({ message : `${msg}: ${res}`, event_type : EVENT_TYPES.SUCCESS })
   }
 }
 
-export const errorHandlerGenerator = (snackbar, action,msg, cb) => (err) => {
+export const errorHandlerGenerator = (notify, msg, cb) => (err) => {
   if (cb !== undefined) cb(err)
-  snackbar(`${msg}: ${err}`, { variant : "error",
-    action,
-    autoHideDuration : 3000, })
+  notify({ message : `${msg}: ${err}`, event_type : EVENT_TYPES.ERROR })
 }
