@@ -1,7 +1,6 @@
-import { IconButton } from "@material-ui/core";
-import { Cancel } from "@material-ui/icons";
-import { useSnackbar } from "notistack";
-import { iconMedium } from "../css/icons.styles";
+
+import { useNotification } from "../utils/hooks/useNotification";
+import { EVENT_TYPES } from "../lib/event-types";
 
 /**
  *  Show Snackbar when error occurs. Can be used in catch blocks
@@ -9,7 +8,7 @@ import { iconMedium } from "../css/icons.styles";
  *
  */
 function HandleError() {
-  const { enqueueSnackbar, closeSnackbar } = useSnackbar();
+  const { notify } = useNotification()
   /**
     *
     * @param {Object} err
@@ -18,22 +17,7 @@ function HandleError() {
     */
   const errorH = (err, prefixMessage, variant) => {
     console.error("an error occured with severity: ", variant, { err })
-    return enqueueSnackbar(
-      `${prefixMessage}: ${err?.message}`,
-      {
-        variant : variant || "error",
-        autoHideDuration : 8000,
-        preventDuplicate : true,
-        action : (key) => (
-          <IconButton
-            onClick={() => closeSnackbar(key)}
-            color="secondary"
-            style={iconMedium}
-          >
-            <Cancel />
-          </IconButton>
-        )
-      })
+    return notify({ message : `${prefixMessage}: ${err?.message}`, event_type : EVENT_TYPES.ERROR, details : err.toString() })
   }
 
   return errorH
