@@ -1,16 +1,13 @@
 //@ts-check
 import React, { useState, useEffect } from "react";
 import { connect } from "react-redux";
-import { withSnackbar } from "notistack";
 import { updateProgress } from "../../lib/store";
 import { bindActionCreators } from "redux";
 import {
   Button, Grid, Paper, Typography
 } from "@material-ui/core";
-import IconButton from "@material-ui/core/IconButton";
 import { useTheme, withStyles } from "@material-ui/core/styles";
 import useMediaQuery from '@material-ui/core/useMediaQuery';
-import CloseIcon from "@material-ui/icons/Close";
 import { withRouter } from "next/router";
 import MesheryMetrics from "../MesheryMetrics";
 import PerformanceCalendar from "./PerformanceCalendar";
@@ -18,7 +15,8 @@ import GenericModal from "../GenericModal";
 import MesheryPerformanceComponent from "./index";
 import fetchPerformanceProfiles from "../graphql/queries/PerformanceProfilesQuery";
 import fetchAllResults from "../graphql/queries/FetchAllResultsQuery";
-import { iconMedium } from "../../css/icons.styles";
+import { useNotification } from "../../utils/hooks/useNotification";
+import { EVENT_TYPES } from "../../lib/event-types";
 
 // const MESHERY_PERFORMANCE_URL = "/api/user/performance/profiles";
 // const MESHERY_PERFORMANCE_TEST_URL = "/api/user/performance/profiles/results";
@@ -75,10 +73,11 @@ const styles = (theme) => ({
   },
 });
 
-function Dashboard({ updateProgress, enqueueSnackbar, closeSnackbar, grafana, router, classes }) {
+function Dashboard({ updateProgress, grafana, router, classes }) {
   const [profiles, setProfiles] = useState({ count : 0, profiles : [] });
   const [tests, setTests] = useState({ count : 0, tests : [] });
   const [runTest, setRunTest] = useState(false);
+  const { notify } = useNotification();
 
   const theme = useTheme();
   const matches = useMediaQuery(theme.breakpoints.up("xs"))
@@ -157,18 +156,7 @@ function Dashboard({ updateProgress, enqueueSnackbar, closeSnackbar, grafana, ro
   function handleError(msg) {
     return function (error) {
       updateProgress({ showProgress : false });
-
-      enqueueSnackbar(`${msg}: ${error}`, {
-        variant : "error",
-        action : function Action(key) {
-          return (
-            <IconButton style={iconMedium} key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
-              <CloseIcon style={iconMedium}/>
-            </IconButton>
-          );
-        },
-        autoHideDuration : 8000,
-      });
+      notify({ message : `${msg}: ${error}`, event_type : EVENT_TYPES.ERROR, details : error.toString() })
     };
   }
 
@@ -252,4 +240,4 @@ const mapStateToProps = (st) => {
 
 const mapDispatchToProps = (dispatch) => ({ updateProgress : bindActionCreators(updateProgress, dispatch), });
 
-export default withStyles(styles)(connect(mapStateToProps, mapDispatchToProps)(withRouter(withSnackbar(Dashboard))));
+export default withStyles(styles)(connect(mapStateToProps, mapDispatchToProps)(withRouter(Dashboard)));
