@@ -8,8 +8,6 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import MUIDataTable from 'mui-datatables';
 import Moment from 'react-moment';
-import { withSnackbar } from 'notistack';
-import CloseIcon from '@material-ui/icons/Close';
 import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
 import { updateResultsSelection, clearResultsSelection, updateProgress, } from '../lib/store';
 import TableSortLabel from '@material-ui/core/TableSortLabel'
@@ -18,6 +16,8 @@ import CustomToolbarSelect from './CustomToolbarSelect';
 import MesheryChart from './MesheryChart';
 import GrafanaCustomCharts from './telemetry/grafana/GrafanaCustomCharts';
 import MesheryResultDialog from './MesheryResultDialog';
+import { withNotify } from '../utils/hooks/useNotification';
+import { EVENT_TYPES } from '../lib/event-types';
 
 
 const styles = (theme) => ({ grid : { padding : theme.spacing(2), },
@@ -99,19 +99,8 @@ class MesheryResults extends Component {
   handleError = (error) => {
     this.props.updateProgress({ showProgress : false });
     // console.log(`error fetching results: ${error}`);
-    const self = this;
-    this.props.enqueueSnackbar(`There was an error fetching results: ${error}`, { variant : 'error',
-      action : (key) => (
-        <IconButton
-          key="close"
-          aria-label="Close"
-          color="inherit"
-          onClick={() => self.props.closeSnackbar(key)}
-        >
-          <CloseIcon />
-        </IconButton>
-      ),
-      autoHideDuration : 8000, });
+    const notify = this.props.notify;
+    notify({ message : `There was an error fetching results: ${error}`, event_type : EVENT_TYPES.ERROR, details : error.toString() });
   }
 
   resetSelectedRowData() {
@@ -463,4 +452,4 @@ const mapStateToProps = (state) => {
 export default withStyles(styles)(connect(
   mapStateToProps,
   mapDispatchToProps,
-)(withSnackbar(MesheryResults)));
+)(withNotify(MesheryResults)));
