@@ -3,7 +3,8 @@ import jsYaml from "js-yaml";
 // eslint-disable-next-line no-unused-vars
 import * as Types from "./types";
 import { promisifiedDataFetch } from "../../../../lib/data-fetch";
-import { useSnackbar } from "notistack";
+import { useNotification } from "../../../../utils/hooks/useNotification";
+import { EVENT_TYPES } from "../../../../lib/event-types";
 
 export default function useDesignLifecycle() {
   const [designName, setDesignName] = useState("Unitled Design")
@@ -13,7 +14,7 @@ export default function useDesignLifecycle() {
     services : {}
   })
   const [designYaml, setDesignyaml] = useState("");
-  const { enqueueSnackbar } = useSnackbar();
+  const { notify } = useNotification()
 
 
   useEffect(function updateDesignYamlFromJson() {
@@ -72,9 +73,9 @@ export default function useDesignLifecycle() {
       method : "POST"
     }).then(data => {
       setDesignId(data[0].id);
-      enqueueSnackbar(`"${designName}" saved successfully`, { variant : "success" })
-    }).catch(() => {
-      enqueueSnackbar("failed to save design file", { variant : "error" })
+      notify({ message : `"${designName}" saved successfully`, event_type : EVENT_TYPES.SUCCESS })
+    }).catch((err) => {
+      notify({ message : `failed to save design file`, event_type : EVENT_TYPES.ERROR, details : err.toString() })
     })
   }
 
@@ -89,22 +90,18 @@ export default function useDesignLifecycle() {
       }),
       method : "POST"
     }).then(() => {
-      enqueueSnackbar(`"${designName}" updated successfully`, { variant : "success" })
-    }).catch(() => {
-      enqueueSnackbar(`couldn't update "${designName}"`, { variant : "error" })
+      notify({ message : `"${designName}" updated successfully`, event_type : EVENT_TYPES.SUCCESS })
+    }).catch((err) => {
+      notify({ message : `failed to update design file`, event_type : EVENT_TYPES.ERROR, details : err.toString() })
     })
   }
 
   function designDelete() {
     return promisifiedDataFetch("/api/pattern/" + designId, { method : "DELETE" }).then(() => {
-      enqueueSnackbar(`Design "${designName}" Deleted`, { variant : "success" })
-      setDesignJson({
-        name : "Unitled Design",
-        services : {}
-      })
+      notify({ message : `Design "${designName}" Deleted`, event_type : EVENT_TYPES.SUCCESS })
       setDesignId(undefined)
       setDesignName("Unitled Design")
-    }).catch(() => enqueueSnackbar("error deleting design", { variant : "error" })
+    }).catch((err) => notify({ message : `failed to delete design file`, event_type : EVENT_TYPES.ERROR, details : err.toString() })
     )
   }
 
