@@ -1,5 +1,5 @@
 /**
- * @generated SignedSource<<e5329bb8772403c10972129fbe02be88>>
+ * @generated SignedSource<<a052580bbfdf42746bd5858d04c0bbd7>>
  * @lightSyntaxTransform
  * @nogrep
  */
@@ -136,6 +136,13 @@ v1 = [
             "kind": "ScalarField",
             "name": "version",
             "storageKey": null
+          },
+          {
+            "alias": null,
+            "args": null,
+            "kind": "ScalarField",
+            "name": "connection_id",
+            "storageKey": null
           }
         ],
         "storageKey": null
@@ -162,16 +169,16 @@ return {
     "selections": (v1/*: any*/)
   },
   "params": {
-    "cacheID": "7e7d2789e340a7c3fd56badc433c0ffc",
+    "cacheID": "7b4896d5a8addeba7cc68215739844e0",
     "id": null,
     "metadata": {},
     "name": "K8sContextSubscription",
     "operationKind": "subscription",
-    "text": "subscription K8sContextSubscription(\n  $selector: PageFilter!\n) {\n  k8sContext: subscribeK8sContext(selector: $selector) {\n    total_count\n    contexts {\n      id\n      name\n      auth\n      cluster\n      server\n      owner\n      created_by\n      meshery_instance_id\n      kubernetes_server_id\n      deployment_type\n      updated_at\n      created_at\n      version\n    }\n  }\n}\n"
+    "text": "subscription K8sContextSubscription(\n  $selector: PageFilter!\n) {\n  k8sContext: subscribeK8sContext(selector: $selector) {\n    total_count\n    contexts {\n      id\n      name\n      auth\n      cluster\n      server\n      owner\n      created_by\n      meshery_instance_id\n      kubernetes_server_id\n      deployment_type\n      updated_at\n      created_at\n      version\n      connection_id\n    }\n  }\n}\n"
   }
 };
 })();
 
-node.hash = "c905518498312d5141459c30dd8d0e41";
+node.hash = "5b1836d53a9593c1301403511170bb49";
 
 module.exports = node;
